Refactor utils into mw_utils package and improve app setup
- Split the monolithic utils.py into a modular mw_utils package with dedicated modules for environment, config, parameter encoding, authentication, HTTP client, and handler creation.
- Updated app.py to use the new utilities, added logging configuration, improved CORS setup, and introduced request ID and health check endpoints.
- Replaced .env.example with a more detailed .env.template.
