--- .env.example
@@ -1,9 +1,0 @@
-# Example environment variables for MoodlewareAPI
-# Copy this file to ".env" and adjust values.
-
-# Port the app listens on (used by compose)
-PORT=8000
-
-# Base URL of your Moodle instance (required)
-MOODLE_URL=https://moodle.school.edu
-
