+++ .env.template
@@ -1,0 +1,20 @@
+################################################
+# MoodlewareAPI environment template
+# Copy to .env and adjust values
+################################################
+
+# Base Moodle instance URL
+# Example: https://moodle.school.edu
+MOODLE_URL=
+
+# Port exposed by the API (compose uses this)
+PORT=8000
+
+# CORS: comma-separated list of allowed origins
+# If empty or "*", all origins are allowed (no credentials)
+ALLOW_ORIGINS=*
+
+# Log level for application
+# Valid: critical,error,warning,info,debug
+# Default when unset: info
+LOG_LEVEL=info
