+++ src/mw_utils/http_client.py
@@ -1,0 +1,5 @@
+from typing import Dict
+
+DEFAULT_HEADERS: Dict[str, str] = {
+    "Accept": "application/json, text/plain;q=0.9, */*;q=0.8"
+}
